backend/requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
PyJWT==2.8.0
python-socketio==5.10.0
eventlet==0.34.2
psycopg[binary]>=3.1.0
python-dotenv==1.0.0
bcrypt==4.1.2
